pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-trading-bot"
version = "0.1.0"
description = "Modular crypto trading bot with grid strategy"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Development Team"}
]
keywords = ["crypto", "trading", "bot", "binance", "ccxt", "grid"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "ccxt>=4.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "structlog>=24.0.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "orjson>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "pytest-playwright>=0.5.0",
    "faker>=28.0.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "bandit>=1.7.0",
    "safety>=2.3.0",
    "pre-commit>=3.5.0",
]
secrets = [
    "keyring>=24.0.0",
]
backtesting = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]
dashboard = [
    "nicegui>=3.4.0",
    "httpx>=0.27.0",
    "plotly>=5.24.0",
]

[project.scripts]
crypto-bot = "crypto_bot.main:cli"
crypto-bot-secrets = "crypto_bot.utils.secrets:secrets_cli"
crypto-bot-security = "crypto_bot.utils.security_check:security_check_cli"
crypto-bot-audit = "crypto_bot.utils.audit:verify_audit_cli"
dashboard = "dashboard.main:main"

[project.urls]
Repository = "https://github.com/yourusername/crypto-trading-bot"

[tool.setuptools.packages.find]
where = ["src", "."]
include = ["crypto_bot*", "dashboard*"]

[tool.setuptools.package-data]
crypto_bot = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["crypto_bot", "dashboard"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "ccxt.*",
    "aiosqlite.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "e2e: marks tests as end-to-end browser tests (deselect with '-m \"not e2e\"')",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/crypto_bot"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
